# docs/openapi/certeus.v1.yaml
...
openapi: 3.0.3
info:
  title: CERTEUS API
  version: "1.0.0"
servers:
  - url: http://localhost:8081
...
paths:
...
  /v1/sources/cache:
    post:
      summary: Cache a legal source by URI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uri]
              properties:
                uri: { type: string }
      responses:
        "200":
          description: Cached source
          content:
            application/json:
              schema:
                type: object
                properties:
                  uri: { type: string }
                  digest: { type: string }
                  path: { type: string }
                  retrieved_at: { type: string }
...